@@ -1,189 +1,179 @@
 {
-  "name": "web",
-  "version": "0.0.0",
+  "name": "react-dashboard-new",
+  "version": "0.1.0",
   "private": true,
-  "engines": {
-    "node": ">=8.10.0"
-  },
-  "browserslist": [
-    ">1%",
-    "last 4 versions",
-    "Firefox ESR",
-    "not ie < 9"
-  ],
   "dependencies": {
-    "@babel/polyfill": "7.4.4",
+    "@babel/core": "7.6.0",
+    "@svgr/webpack": "4.3.2",
+    "@typescript-eslint/eslint-plugin": "^2.2.0",
+    "@typescript-eslint/parser": "^2.2.0",
     "awesome-bootstrap-checkbox": "^1.0.1",
-    "bluebird": "^3.5.5",
-    "body-parser": "^1.19.0",
-    "bootstrap": "4.3.1",
-    "bundle-loader": "^0.5.6",
+    "axios": "^0.19.0",
+    "babel-eslint": "10.0.3",
+    "babel-jest": "^24.9.0",
+    "babel-loader": "8.0.6",
+    "babel-plugin-named-asset-import": "^0.3.4",
+    "babel-preset-react-app": "^9.0.2",
+    "bootstrap": "^4.3.1",
+    "camelcase": "^5.2.0",
+    "case-sensitive-paths-webpack-plugin": "2.2.0",
     "classnames": "^2.2.6",
     "cookie-parser": "^1.4.4",
-    "core-js": "^3.2.1",
-    "express": "^4.17.1",
+    "css-loader": "2.1.1",
+    "dotenv": "^8.2.0",
+    "dotenv-expand": "5.1.0",
+    "eslint": "^6.1.0",
+    "eslint-config-react-app": "^5.0.2",
+    "eslint-loader": "3.0.2",
+    "eslint-plugin-flowtype": "3.13.0",
+    "eslint-plugin-import": "2.18.2",
+    "eslint-plugin-jsx-a11y": "6.2.3",
+    "eslint-plugin-react": "7.14.3",
+    "eslint-plugin-react-hooks": "^1.6.1",
     "express-graphql": "^0.9.0",
     "express-jwt": "^5.3.1",
+    "file-loader": "3.0.1",
     "font-awesome": "^4.7.0",
+    "fs-extra": "7.0.1",
     "glyphicons-halflings": "^1.9.1",
-    "graphql": "^14.4.2",
+    "graphql": "^14.5.8",
     "graphql-relay": "^0.6.0",
     "graphql-sequelize": "^9.3.6",
-    "history": "^4.9.0",
-    "imports-loader": "0.8.0",
-    "isomorphic-fetch": "^2.2.1",
-    "isomorphic-style-loader": "^5.1.0",
-    "jquery": "3.4.1",
+    "helmet": "^3.21.2",
+    "html-webpack-plugin": "4.0.0-beta.5",
+    "identity-obj-proxy": "3.0.0",
+    "is-wsl": "^1.1.0",
+    "jest": "24.9.0",
+    "jest-environment-jsdom-fourteen": "0.1.0",
+    "jest-resolve": "24.9.0",
+    "jest-watch-typeahead": "0.4.0",
     "jsonwebtoken": "^8.5.1",
-    "messenger": "git+https://github.com/HubSpot/messenger.git#v1.4.2",
-    "node-fetch": "^2.6.0",
+    "mini-css-extract-plugin": "0.8.0",
+    "mysql2": "^2.0.0",
+    "node-sass": "^4.13.0",
     "normalize.css": "^8.0.1",
-    "passport": "^0.4.0",
-    "passport-facebook": "^3.0.0",
-    "pretty-error": "^2.1.1",
+    "npm-run-all": "^4.1.5",
+    "optimize-css-assets-webpack-plugin": "5.0.3",
+    "pnp-webpack-plugin": "1.5.0",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-normalize": "7.0.1",
+    "postcss-preset-env": "6.7.0",
+    "postcss-safe-parser": "4.0.1",
     "prop-types": "^15.7.2",
-    "query-string": "^6.8.2",
-    "react": "16.9.0",
-    "react-cookie": "^4.0.1",
-    "react-dom": "16.9.0",
+    "react": "^16.11.0",
+    "react-app-polyfill": "^1.0.4",
+    "react-dev-utils": "^9.1.0",
+    "react-dom": "^16.11.0",
     "react-google-maps": "^9.4.5",
-    "react-redux": "^7.1.0",
-    "react-router": "5.0.1",
-    "react-router-dom": "5.0.1",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
     "react-sparklines": "^1.7.0",
-    "reactstrap": "8.0.1",
-    "recharts": "^1.7.1",
+    "react-toastify": "^5.4.1",
+    "reactstrap": "^8.1.1",
+    "recharts": "^1.8.5",
     "redux": "^4.0.4",
-    "redux-logger": "^3.0.1",
     "redux-thunk": "^2.3.0",
-    "sequelize": "^5.15.1",
-    "serialize-javascript": "^1.7.0",
-    "source-map-support": "^0.5.13",
-    "sqlite3": "^4.0.9",
-    "whatwg-fetch": "^3.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "7.5.5",
-    "@babel/node": "7.5.5",
-    "@babel/plugin-transform-react-constant-elements": "7.5.0",
-    "@babel/plugin-transform-react-inline-elements": "7.2.0",
-    "@babel/preset-env": "7.5.5",
-    "@babel/preset-flow": "7.0.0",
-    "@babel/preset-react": "7.0.0",
-    "@babel/preset-stage-2": "7.0.0",
-    "assets-webpack-plugin": "^3.9.10",
-    "autoprefixer": "^9.6.1",
-    "babel-core": "^7.0.0-0",
-    "babel-eslint": "^10.0.2",
-    "babel-jest": "^24.8.0",
-    "babel-loader": "8.0.6",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "browser-sync": "^2.26.7",
-    "chokidar": "^3.0.2",
-    "cross-env": "^5.2.0",
-    "css-loader": "^3.2.0",
-    "enzyme": "^3.10.0",
-    "eslint": "^6.1.0",
-    "eslint-config-airbnb": "^18.0.0",
-    "eslint-config-prettier": "^6.0.0",
-    "eslint-import-resolver-node": "^0.3.2",
-    "eslint-loader": "^2.2.1",
-    "eslint-plugin-css-modules": "^2.11.0",
-    "eslint-plugin-flowtype": "^4.2.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.0",
-    "eslint-plugin-react": "^7.14.3",
-    "file-loader": "^4.2.0",
-    "flow-bin": "^0.105.2",
-    "front-matter": "^3.0.2",
-    "glob": "^7.1.4",
-    "husky": "^3.0.3",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.8.0",
-    "jscodeshift": "^0.6.4",
-    "lint-staged": "^9.2.1",
-    "markdown-it": "^9.1.0",
-    "mkdirp": "^0.5.1",
-    "node-sass": "4.12.0",
-    "null-loader": "^3.0.0",
-    "pixrem": "^5.0.0",
-    "pleeease-filters": "^4.0.0",
-    "postcss": "^7.0.17",
-    "postcss-calc": "^7.0.1",
-    "postcss-color-function": "^4.1.0",
-    "postcss-custom-media": "^7.0.8",
-    "postcss-custom-properties": "^9.0.2",
-    "postcss-custom-selectors": "^5.1.2",
-    "postcss-flexbugs-fixes": "^4.1.0",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-media-minmax": "^4.0.0",
-    "postcss-nested": "^4.1.2",
-    "postcss-nesting": "^7.0.1",
-    "postcss-pseudoelements": "^5.0.0",
-    "postcss-selector-matches": "^4.0.0",
-    "postcss-selector-not": "^4.0.0",
-    "prettier": "^1.18.2",
-    "raw-loader": "^3.1.0",
-    "react-deep-force-update": "^2.1.3",
-    "react-dev-utils": "^9.0.2",
-    "react-error-overlay": "^6.0.0",
-    "react-hot-loader": "^4.12.11",
-    "react-test-renderer": "^16.9.0",
-    "redux-mock-store": "^1.5.3",
-    "rimraf": "^2.6.3",
+    "resolve": "1.12.0",
+    "resolve-url-loader": "2",
     "sass-loader": "7.2.0",
-    "sinon": "^7.4.1",
-    "stylelint": "^10.1.0",
-    "stylelint-config-standard": "^18.3.0",
-    "stylelint-order": "^3.0.1",
-    "svg-url-loader": "^3.0.0",
-    "url-loader": "^2.1.0",
-    "webpack": "^4.39.1",
-    "webpack-bundle-analyzer": "^3.4.1",
-    "webpack-dev-middleware": "^3.7.0",
-    "webpack-hot-middleware": "^2.25.0",
-    "webpack-node-externals": "^1.7.2"
-  },
-  "resolutions": {
-    "jquery": "3.3.1"
-  },
-  "lint-staged": {
-    "*.{js,jsx}": [
-      "eslint --fix",
-      "git add --force"
-    ],
-    "*.{json,graphql}": [
-      "prettier --write",
-      "git add --force"
-    ],
-    "*.{css,less,scss,sss}": [
-      "stylelint --fix",
-      "git add --force"
-    ]
+    "semver": "6.3.0",
+    "sequelize": "^5.21.2",
+    "sqlite3": "^4.1.0",
+    "style-loader": "1.0.0",
+    "terser-webpack-plugin": "1.4.1",
+    "ts-pnp": "1.1.4",
+    "url-loader": "2.1.0",
+    "uuid": "^3.3.3",
+    "webpack": "4.41.0",
+    "webpack-dev-server": "3.2.1",
+    "webpack-manifest-plugin": "2.1.1",
+    "workbox-webpack-plugin": "4.3.1"
   },
   "scripts": {
-    "precommit": "lint-staged",
-    "lint-js": "eslint --ignore-path .gitignore --ignore-pattern \"!**/.*\" .",
-    "lint-css": "stylelint \"src/**/*.{css,less,scss,sss}\"",
-    "lint": "yarn run lint-js && yarn run lint-css",
-    "fix-js": "yarn run lint-js --fix",
-    "fix-css": "yarn run lint-css --fix",
-    "fix": "yarn run fix-js && yarn run fix-css",
-    "check": "flow check",
-    "test": "jest",
-    "test-watch": "yarn run test --watch --notify",
-    "test-cover": "yarn run test --coverage",
-    "coverage": "yarn run test-cover && opn coverage/lcov-report/index.html",
-    "clean": "babel-node tools/run clean",
-    "copy": "babel-node tools/run copy",
-    "bundle": "babel-node tools/run bundle",
-    "build": "babel-node tools/run build",
-    "build-stats": "yarn run build --release --analyse",
-    "deploy": "babel-node tools/run deploy",
-    "render": "babel-node tools/run render",
-    "serve": "babel-node tools/run runServer",
-    "start": "babel-node tools/run start"
+    "start": "node scripts/start.js",
+    "build": "cross-env PUBLIC_URL='/react-dashboard' node scripts/build.js",
+    "test": "node scripts/test.js",
+    "server": "node server/start.js",
+    "dev": "cross-env NODE_ENV=development run-p server start"
+  },
+  "proxy": "http://localhost:5000",
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "jest": {
+    "roots": [
+      "<rootDir>/src"
+    ],
+    "collectCoverageFrom": [
+      "src/**/*.{js,jsx,ts,tsx}",
+      "!src/**/*.d.ts"
+    ],
+    "setupFiles": [
+      "react-app-polyfill/jsdom"
+    ],
+    "setupFilesAfterEnv": [],
+    "testMatch": [
+      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
+      "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
+    ],
+    "testEnvironment": "jest-environment-jsdom-fourteen",
+    "transform": {
+      "^.+\\.(js|jsx|ts|tsx)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
+      "^(?!.*\\.(js|jsx|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
+    },
+    "transformIgnorePatterns": [
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|ts|tsx)$",
+      "^.+\\.module\\.(css|sass|scss)$"
+    ],
+    "modulePaths": [],
+    "moduleNameMapper": {
+      "^react-native$": "react-native-web",
+      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
+    },
+    "moduleFileExtensions": [
+      "web.js",
+      "js",
+      "web.ts",
+      "ts",
+      "web.tsx",
+      "tsx",
+      "json",
+      "web.jsx",
+      "jsx",
+      "node"
+    ],
+    "watchPlugins": [
+      "jest-watch-typeahead/filename",
+      "jest-watch-typeahead/testname"
+    ]
+  },
+  "babel": {
+    "presets": [
+      "react-app"
+    ]
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.7.0",
+    "@babel/node": "^7.7.0",
+    "@babel/plugin-proposal-class-properties": "^7.7.0",
+    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.7.1",
+    "cross-env": "^6.0.3"
   }
 }
