@@ -1,24 +1,23 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import {withRouter, Link} from 'react-router-dom';
 
 import Icon from '../Icon';
 import LinksGroup from './LinksGroup/LinksGroup';
 
-import s from './Sidebar.scss';
+import s from './Sidebar.module.scss';
 
 const Sidebar = () => (
   <nav className={s.root}>
     <header className={s.logo}>
-      <Link to="/app">
+      <Link to="/app/main">
         <Icon glyph="logo" />
       </Link>
     </header>
     <ul className={s.nav}>
       <LinksGroup
         header="Dashboard"
-        headerLink="/app"
+        headerLink="/app/main"
         glyph="dashboard"
       />
       <LinksGroup
@@ -70,4 +69,4 @@ function mapStateToProps(store) {
   };
 }
 
-export default withRouter(connect(mapStateToProps)(withStyles(s)(Sidebar)));
+export default withRouter(connect(mapStateToProps)(Sidebar));
