@@ -1,17 +1,6 @@
-/**
- * Flatlogic Dashboards (https://flatlogic.com/admin-dashboards)
- *
- * Copyright © 2015-present Flatlogic, LLC. All rights reserved.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE.txt file in the root directory of this source tree.
- */
-
 import React from 'react';
 import PropTypes from 'prop-types';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 
-import s from './ErrorPage.scss';
 
 class ErrorPage extends React.Component {
   static propTypes = {
@@ -23,17 +12,6 @@ class ErrorPage extends React.Component {
   };
 
   render() {
-    if (__DEV__) {
-      const {error} = this.props;
-      return (
-        <div>
-          <h1>{error.name}</h1>
-          <p>{error.message}</p>
-          <pre>{error.stack}</pre>
-        </div>
-      );
-    }
-
     return (
       <div>
         <h1>Error</h1>
@@ -43,5 +21,4 @@ class ErrorPage extends React.Component {
   }
 }
 
-export {ErrorPage as ErrorPageWithoutStyle};
-export default withStyles(s)(ErrorPage);
+export default ErrorPage;
