@@ -1,161 +1,81 @@
 import React from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import {
-  Row, Col, Button, Breadcrumb, BreadcrumbItem
+  Row, Col, Button,
 } from 'reactstrap';
-/* eslint-disable */
-import 'imports-loader?$=jquery,this=>window!jquery';
-import 'imports-loader?$=jquery,this=>window!messenger/build/js/messenger';
-/* eslint-enable */
 
+import { toast } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import uuid from 'uuid/v4'
 import Widget from '../../components/Widget';
-
-import s from './Notifications.scss';
-
-// todo @franckeeva what about server side rendering? this will fail unless launched as lazy route
-const {Messenger} = window;
-
-/* eslint-disable */
-function initializationMessengerCode() {
-  (function () {
-    let $,
-      FlatMessage,
-      spinner_template,
-      __hasProp = {}.hasOwnProperty,
-      __extends = function (child, parent) {
-        for (const key in parent) {
-          if (__hasProp.call(parent, key)) child[key] = parent[key];
-        }
-
-        function ctor() {
-          this.constructor = child;
-        }
-
-        ctor.prototype = parent.prototype;
-        child.prototype = new ctor();
-        child.__super__ = parent.prototype;
-        return child;
-      };
-
-    $ = jQuery;
-
-    spinner_template = '<div class="messenger-spinner">\n    <span class="messenger-spinner-side messenger-spinner-side-left">\n        <span class="messenger-spinner-fill"></span>\n    </span>\n    <span class="messenger-spinner-side messenger-spinner-side-right">\n        <span class="messenger-spinner-fill"></span>\n    </span>\n</div>';
-
-    FlatMessage = (function (_super) {
-      __extends(FlatMessage, _super);
-
-      function FlatMessage() {
-        return FlatMessage.__super__.constructor.apply(this, arguments);
-      }
-
-      FlatMessage.prototype.template = function (opts) {
-        let $message;
-        $message = FlatMessage.__super__.template.apply(this, arguments);
-        $message.append($(spinner_template));
-        return $message;
-      };
-
-      return FlatMessage;
-    }(Messenger.Message));
-
-    Messenger.themes.air = {
-      Message: FlatMessage,
-    };
-  }).call(window);
-}
-/* eslint-enable */
+import s from './Notifications.module.scss';
 
 class Notifications extends React.Component {
-  constructor() {
-    super();
-    this.addSuccessNotification = this.addSuccessNotification.bind(this);
-    this.addInfoNotification = this.addInfoNotification.bind(this);
-    this.addErrorNotification = this.addErrorNotification.bind(this);
-    this.toggleLocation = this.toggleLocation.bind(this);
-    this.state = {
-      locationClasses: 'messenger-fixed messenger-on-bottom messenger-on-right',
-    };
+  
+  state = {
+    options: {
+      position: "top-right",
+      autoClose: 5000,
+      closeOnClick: false,
+      pauseOnHover: false,
+      draggable: true
+    }
   }
-
 
   componentDidMount() {
-    initializationMessengerCode();
-    Messenger.options = {
-      extraClasses: this.state.locationClasses,
-      theme: 'air',
-    };
-    Messenger().post('Thanks for checking out Messenger!');
-  }
-
-  addSuccessNotification() {
-    Messenger().post({
-      extraClasses: this.state.locationClasses,
-      message: 'Showing success message was successful!',
-      type: 'success',
-      showCloseButton: true,
-    });
-    return false;
-  }
-
-  addInfoNotification() {
-    const msg = Messenger().post({
-      extraClasses: this.state.locationClasses,
-      message: 'Launching thermonuclear war...',
-      actions: {
-        cancel: {
-          label: 'cancel launch',
-          action: () => msg.update({
-            message: 'Thermonuclear war averted', type: 'success', actions: false,
-          }),
-        },
-      },
-    });
-
-    return false;
-  }
-
-  addErrorNotification() {
-    let i = 0;
-    Messenger().run({
-      errorMessage: 'Error destroying alien planet',
-      successMessage: 'Alien planet destroyed!',
-      extraClasses: this.state.locationClasses,
-      action(opts) {
-        /* eslint-disable */
-        if (++i < 3) {
-          return opts.error({
-            status: 500,
-            readyState: 0,
-            responseText: 0,
-          });
-        }
-        /* eslint-enable */
-        return opts.success();
-      },
+    toast.success('Thanks for checking out Messenger!', {
+      position: "bottom-right",
+      autoClose: 5000,
+      closeOnClick: true,
+      pauseOnHover: false,
+      draggable: true
     });
   }
 
-  toggleLocation(vertical = 'top', horizontal = '') {
-    let className = `messenger-fixed messenger-on-${vertical}`;
-    className += (horizontal === '') ? '' : ` messenger-on-${horizontal}`;
-    this.setState({
-      locationClasses: className,
-    });
-    Messenger.options = {
-      extraClasses: className,
-      theme: 'air',
-    };
-    Messenger();
-  }
+  addSuccessNotification = () => toast.success('Showing success message was successful!', this.state.options);
 
+  toggleLocation = (location) => {
+    this.setState(prevState => ({
+      options: {
+        ...prevState.options,
+        position: location
+      }
+    }));
+  }
+ 
+  addInfoNotification = () => {
+    let id = uuid();
+    toast.info( 
+    <div className="d-flex flex-column align-items-center">
+      Launching thermonuclear war...
+      <Button onClick={() => this.launchNotification(id)} outline color="default" size="sm" className="width-100 mb-xs mr-xs mt-1 ml-4">Cancel launch</Button>
+    </div>, 
+    {...this.state.options,toastId: id},
+    );
+  }
+ 
+  launchNotification = (id) => toast.update(id, { ...this.state.options, render: "Thermonuclear war averted", type: toast.TYPE.SUCCESS });
+ 
+  addErrorNotification = () => {
+    let id = uuid();
+    toast.error(
+    <div className="d-flex flex-column align-items-center">
+      Error destroying alien planet
+      <Button onClick={() => this.retryNotification(id)} outline color="default" size="sm" className="width-100 mb-xs mr-xs mt-1 ml-4">Retry</Button>
+    </div>, 
+    {...this.state.options,toastId: id}
+    );
+  } 
+
+  retryNotification = (id) =>  toast.update(id, {...this.state.options, render: 'Alien planet destroyed!', type: toast.TYPE.SUCCESS });
+  
   render() {
     return (
       <div className={s.root}>
-        <Breadcrumb>
-          <BreadcrumbItem>YOU ARE HERE</BreadcrumbItem>
-          <BreadcrumbItem active>UI Notifications</BreadcrumbItem>
-        </Breadcrumb>
-        <h1 className="page-title mb-lg">Messages - <span className="fw-semi-bold">Notifications</span>
+        <ol className="breadcrumb">
+          <li className="breadcrumb-item">YOU ARE HERE</li>
+          <li className="breadcrumb-item active">UI Notifications</li>
+        </ol>
+        <h1 className="page-title">Messages - <span className="fw-semi-bold">Notifications</span>
         </h1>
 
         <Widget title={<h6> Messenger </h6>} close collapse settings>
@@ -166,38 +86,36 @@ class Notifications extends React.Component {
                 them
                 to change notifications position:</p>
               <div className="location-selector">
-                { /* eslint-disable */}
                 <div
                   className="bit top left" onClick={() => {
-                    this.toggleLocation('top', 'left');
+                    this.toggleLocation('top-left');
                   }}
                 />
                 <div
                   className="bit top right" onClick={() => {
-                    this.toggleLocation('top', 'right');
+                    this.toggleLocation('top-right');
                   }}
                 />
                 <div
                   className="bit top" onClick={() => {
-                    this.toggleLocation('top', '');
+                    this.toggleLocation('top-center');
                   }}
                 />
                 <div
                   className="bit bottom left" onClick={() => {
-                    this.toggleLocation('bottom', 'left');
+                    this.toggleLocation('bottom-left');
                   }}
                 />
                 <div
                   className="bit bottom right" onClick={() => {
-                    this.toggleLocation('bottom', 'right');
+                    this.toggleLocation('bottom-right');
                   }}
                 />
                 <div
                   className="bit bottom" onClick={() => {
-                    this.toggleLocation('bottom', '');
+                    this.toggleLocation('bottom-center');
                   }}
                 />
-                { /* eslint-enable */}
               </div>
             </Col>
 
@@ -219,10 +137,10 @@ class Notifications extends React.Component {
               <h5 className="m-t-1">Dead Simple Usage</h5>
               <p>Just few lines of code to instantiate a notifications object. Does not require
                 passing any options:</p>
-              <pre><code>Messenger().post(&quot;Thanks for checking out Messenger!&quot;);</code></pre>
+              <pre><code>{'toast("Thanks for checking out Messenger!");'}</code></pre>
               <p>More complex example:</p>
               <pre>
-                <code>{'\nMessenger().post({\n  message: \'There was an explosion while processing your request.\',\n  type: \'error\',\n  showCloseButton: true\n});\n\n'}
+                <code>{'\ntoast.success( \'There was an explosion while processing your request.\', { \n position: location,\n autoClose: 5000, \n hideProgressBar: false, \n closeOnClick: true,\n pauseOnHover: true, \n draggable: true \n});\n\n'}
                 </code>
               </pre>
             </Col>
@@ -233,4 +151,4 @@ class Notifications extends React.Component {
   }
 }
 
-export default withStyles(s)(Notifications);
+export default Notifications;
