@@ -1,5 +1,4 @@
 import React, {PureComponent} from 'react';
-import withStyles from 'isomorphic-style-loader/lib/withStyles';
 import {
   Row,
   Col,
@@ -15,10 +14,10 @@ import {
 
 import Widget from '../../components/Widget';
 
-import s from './Profile.scss';
+import s from './Profile.module.scss';
 
 class Profile extends PureComponent {
-  onSubmit(e) { // eslint-disable-line
+  onSubmit(e) { 
     e.preventDefault();
   }
 
@@ -42,15 +41,15 @@ class Profile extends PureComponent {
               <Form onSubmit={this.onSubmit}>
                 <FormGroup>
                   <Label for="input-name">Name</Label>
-                  <Input size="lg" type="text" name="name" id="input-name"/>
+                  <Input bsSize="lg" type="text" name="name" id="input-name"/>
                 </FormGroup>
                 <FormGroup>
                   <Label for="input-email">Email</Label>
-                  <Input size="lg" type="email" name="email" id="input-email"/>
+                  <Input bsSize="lg" type="email" name="email" id="input-email"/>
                 </FormGroup>
                 <FormGroup>
                   <Label for="input-password">Password</Label>
-                  <Input size="lg" type="password" name="password" id="input-password"/>
+                  <Input bsSize="lg" type="password" name="password" id="input-password"/>
                 </FormGroup>
                 <div className="d-flex justify-content-between align-items-center">
                   <div className="d-flex">
@@ -74,4 +73,4 @@ class Profile extends PureComponent {
   }
 }
 
-export default withStyles(s)(Profile);
+export default Profile;
