@@ -1,44 +1,35 @@
 import {
-  LOGIN_REQUEST,
-  LOGIN_SUCCESS,
-  LOGIN_FAILURE,
-  LOGOUT_SUCCESS,
+  LOGIN_REQUEST, LOGIN_SUCCESS, LOGIN_FAILURE, LOGOUT_SUCCESS,
 } from '../actions/user';
 
-// The auth reducer. The starting state sets authentication
-// based on a token being in local storage. In a real app,
-// we would also want a util to check if the token is expired.
-export default function auth(
-  state = {
-    isFetching: false,
-    isAuthenticated: false,
-  },
-  action,
-) {
+const token = localStorage.getItem('token');
+export default function auth(state = {
+  isFetching: false,
+  isAuthenticated: !!token,
+}, action) {
   switch (action.type) {
-    case LOGIN_REQUEST:
-      return Object.assign({}, state, {
-        isFetching: true,
-        isAuthenticated: false,
-        user: action.creds,
-      });
-    case LOGIN_SUCCESS:
-      return Object.assign({}, state, {
-        isFetching: false,
-        isAuthenticated: true,
-        errorMessage: '',
-      });
-    case LOGIN_FAILURE:
-      return Object.assign({}, state, {
-        isFetching: false,
-        isAuthenticated: false,
-        errorMessage: action.message,
-      });
-    case LOGOUT_SUCCESS:
-      return Object.assign({}, state, {
-        isAuthenticated: false,
-      });
-    default:
-      return state;
+      case LOGIN_REQUEST:
+          return Object.assign({}, state, {
+              isFetching: true,
+              isAuthenticated: false,
+          });
+      case LOGIN_SUCCESS:
+          return Object.assign({}, state, {
+              isFetching: false,
+              isAuthenticated: true,
+              errorMessage: '',
+          });
+      case LOGIN_FAILURE:
+          return Object.assign({}, state, {
+              isFetching: false,
+              isAuthenticated: false,
+              errorMessage: action.payload,
+          });
+      case LOGOUT_SUCCESS:
+          return Object.assign({}, state, {
+              isAuthenticated: false,
+          });
+      default:
+          return state;
   }
 }
